crypto: pass struct crypto_alg directly to alg_test()
algboss/cryptomgr_schedule_test/cryptomgr_test currently pass the
algorithm to test by name -- stringly typed.

This is unfortunate as it opens all kinds of potential issues with race
conditions and ultimately the question of whether we actually tested
what we wanted to test.

I see no reason why we should pass the algorithm by name except for the
fact that the crypto API itself is stringly typed when
requesting/instantiating algorithms. Maybe there should be a way to
instantiate an algorithm by alg pointer too?

In any case, let's pass the alg directly so we know precisely what
algorithm we are actually testing.

Signed-off-by: Vegard Nossum <vegard.nossum@oracle.com>